fix: arreglar código bmi_full.feature y BMIFullSteps.java
Arreglar el código al idioma correcto, @Dado,@Cuando,@Entonces.
Corregido fallos de minúsculas, mayúsculas, comas, puntos para la correcta ejecución de los test con RunCucumberTest.java
