--- java-project-healthcalc/src/test/java/bdd/RunCucumberTest.java
@@ -1,20 +1,0 @@
-package bdd; // Debe coincidir con la carpeta donde está este archivo
-
-import org.junit.platform.suite.api.ConfigurationParameter;
-import org.junit.platform.suite.api.IncludeEngines;
-import org.junit.platform.suite.api.SelectPackages;
-import org.junit.platform.suite.api.Suite;
-
-import static io.cucumber.junit.platform.engine.Constants.GLUE_PROPERTY_NAME;
-import static io.cucumber.junit.platform.engine.Constants.PLUGIN_PROPERTY_NAME;
-
-@Suite
-@IncludeEngines("cucumber")
-// 1. Apuntamos a la carpeta de resources donde está el .feature
-@SelectPackages("")
-// 2. Apuntamos al paquete exacto donde está BMIFullSteps.java
-@ConfigurationParameter(key = GLUE_PROPERTY_NAME, value = "stepdefinitions")
-@ConfigurationParameter(key = PLUGIN_PROPERTY_NAME, value = "pretty, html:target/cucumber-report.html")
-public class RunCucumberTest { 
-} 
-
