+++ java-project-healthcalc/src/test/resources/bmi_full.feature
@@ -1,88 +1,64 @@
+# language: es
 
-Feature: Cálculo del Índice de Masa Corporal de un usuario
+Característica: Cálculo del Índice de Masa Corporal de un usuario
   Como estudiante de Ingeniería de la Salud
   Quiero calcular el Índice de Masa Corporal de un usuario
   Para evaluar la adecuación del peso de una persona en relación con su altura
 
-  Background:
-    Given La calculadora de salud está iniciada
+  Antecedentes:
+    Dado la calculadora de salud está iniciada
 
-  @EdgeCase
-  Scenario: Cálculo del BMI de 75kg y 1.50m
-    Given El peso introducido en kg es 75
-    Given La altura introducida en centímetros es 150
-    When Ejecuto la operación de BMI
-    Then El BMI debe ser 33,33
-    Then El resultado debe ser Obese Class I
+  Escenario: Cálculo del BMI con datos válidos
+    Dado el peso introducido es 75,0
+    Y la altura introducida es 150,0
+    Cuando ejecuto la operación de BMI
+    Entonces el bmi debe ser 33,33
+    Y el resultado debe ser "Obese Class I"
 
-  @EdgeCase
-  Scenario: Cálculo del BMI de 97kg y 1.75m
-    Given El peso introducido en kg es 97
-    Given La altura introducida en centímetros es 175
-    When Ejecuto la operación de BMI
-    Then El BMI debe ser 31,67
-    Then El resultado debe ser Obese Class I
+  Escenario: Cálculo del BMI de 97kg y 1.75m
+    Dado el peso introducido es 97,0
+    Y la altura introducida es 175,0
+    Cuando ejecuto la operación de BMI
+    Entonces el bmi debe ser 31,67
+    Y el resultado debe ser "Obese Class I"
 
-  @EdgeCase
-  Scenario: Cálculo del BMI de 70kg y 1.90m
-    Given El peso introducido en kg es 70
-    Given La altura introducida en centímetros es 190
-    When Ejecuto la operación de BMI
-    Then El BMI debe ser 19,39
-    Then El resultado debe ser Normal
+  Escenario: Cálculo del BMI de 70kg y 1.90m
+    Dado el peso introducido es 70,0
+    Y la altura introducida es 190,0
+    Cuando ejecuto la operación de BMI
+    Entonces el bmi debe ser 19,39
+    Y el resultado debe ser "Normal"
+  
+  @Error
+  Escenario: Intento de cálculo de BMI con números negativos
+    Dado el peso introducido es -70,0
+    Y la altura introducida es -190,0
+    Cuando ejecuto la operación de BMI
+    Entonces el sistema debe lanzar una excepción
 
-  @ErrorHandling
-  Scenario Outline: Intento de cálculo de BMI con números negativos
-    Given El peso introducido en kg es -70
-    Given La altura introducida en centímetros es -190
-    When Ejecuto la operación de BMI
-    Then El sistema debe lanzar una excepción
+  @Error
+  Escenario: Intento de cálculo de BMI con número fuera del rango establecido para el peso
+    Dado el peso introducido es 800,0
+    Y la altura introducida es 190,0
+    Cuando ejecuto la operación de BMI
+    Entonces el sistema debe lanzar una excepción
 
-    Examples:
-      | numero |
-      | -1     |
-      | -2     |
-      | -5     |
-      | -10    |
-      | -50    |
+  @Error
+  Escenario: Intento de cálculo de BMI con número fuera del rango establecido para la altura
+    Dado el peso introducido es 75,0
+    Y la altura introducida es 25,0
+    Cuando ejecuto la operación de BMI
+    Entonces el sistema debe lanzar una excepción
 
-  @ErrorHandling
-  Scenario Outline: Intento de cálculo de BMI con número fuera del rango establecido para el peso
-    Given El peso introducido en kg es 800
-    Given La altura introducida en centímetros es 190
-    When Ejecuto la operación de BMI
-    Then El sistema debe lanzar una excepción
+  Esquema del escenario: Verificación de cálculos exitosos múltiples
+    Dado el peso introducido es <weight>
+    Y la altura introducida es <height>
+    Cuando ejecuto la operación de BMI
+    Entonces el bmi debe ser <bmi>
+    Y el resultado debe ser "<resultado>"
 
-    Examples:
-      | numero |
-      | 701     |
-      | 750     |
-      | 800     |
-
-  @ErrorHandling
-  Scenario Outline: Intento de cálculo de BMI con número fuera del rango establecido para la altura
-    Given El peso introducido en kg es 75
-    Given La altura introducida en centímetros es 25
-    When Ejecuto la operación de BMI
-    Then El sistema debe lanzar una excepción
-
-    Examples:
-      | numero |
-      | 20     |
-      | 29     |
-      | 301     |
-      | 350     |
-      | 400     |
-
-  Scenario Outline: Verificación de cálculos exitosos múltiples
-    Given el peso de entrada es <weight>
-    Given la altura de entrada es <height>
-    When ejecuto la operación de BMI
-    Then el BMI debe ser <bmi>
-    Then el resultado debe ser <resultado>
-
-    Examples:
-      | weight(kg)|   height(cm)   |  BMI  |    resultado     | 
+    Ejemplos:
+      |  weight   |    height      |  bmi  |    resultado     | 
       |   75      |      150       | 33,33 |   Obese Class I  |
       |   97      |      175       | 31,67 |   Obese Class I  |
       |   70      |      190       | 19,39 |      Normal      |
